dataset suite

+++ src/main/scala/week4/Person.scala
@@ -3,6 +3,6 @@ package week4
 /**
   * Created by matijav on 20/03/2017.
   */
-case class Person(id: Int, name: String, age: Int, city: String, country: String) {
+case class Person(id: Long, name: String, age: Long, city: String, country: String) {
 
 }
